+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - fix: The value of the local variable 'onOpenSubcription' isn't used (api_client).
 - fix: cancel subscribe warning.
 - feat: test for jwt util
+- feat: type definitions test
+- fix: payload type conversion
 
 # 0.0.1-nullsafety.0